refactor: move files
* rename folder. @types -> types

* rename filesystem-data to filesystem

* rename models.d.ts -> models.ts

* move view.ts to views folder
