+++ tsconfig.json
@@ -20,6 +20,6 @@
   },
   "include": [
     "**/*.ts",
-    "src/@types",
+    "src/types",
   ]
 }
